requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
PyMySQL==1.1.1
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.10.1
python-multipart==0.0.20
a2wsgi==1.10.8
google-genai==1.34.0
httpx==0.28.1
pytest==8.4.2
